requirements.txt
streamlit
streamlit-webrtc
psycopg2-binary
opencv-python-headless
qrcode
pandas